Criada tela de Membro Selecionado. Corrigido encoding da resposta JSON.

+++ app/src/main/java/br/com/enterprise/pytaco/pojo/Clube.java
@@ -4,8 +4,8 @@ public class Clube extends EntidadeBase {
 
     private String nome;
     private String descricao;
-    private Integer qtdPytaco;
-    private Integer qtdFicha;
+    private Double qtdPytaco;
+    private Double qtdFicha;
     private Usuario usuario;
 
     public String getNome() {
@@ -24,19 +24,19 @@ public class Clube extends EntidadeBase {
         this.descricao = descricao;
     }
 
-    public Integer getQtdPytaco() {
+    public Double getQtdPytaco() {
         return qtdPytaco;
     }
 
-    public void setQtdPytaco(Integer qtdPytaco) {
+    public void setQtdPytaco(Double qtdPytaco) {
         this.qtdPytaco = qtdPytaco;
     }
 
-    public Integer getQtdFicha() {
+    public Double getQtdFicha() {
         return qtdFicha;
     }
 
-    public void setQtdFicha(Integer qtdFicha) {
+    public void setQtdFicha(Double qtdFicha) {
         this.qtdFicha = qtdFicha;
     }
